open Dapp inside Safe
